requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.3
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2023.3
